refactor: modularize app.py routes into Flask Blueprints (#84)
Break the monolithic 1119-line app.py into focused route modules:

- routes/page_routes.py — SPA page routes (/, /posts, /editor, etc.)
- routes/server_routes.py — Hugo server management (/api/server/*)
- routes/post_routes.py — Article listing, cache, references
- routes/file_routes.py — File read/save, post creation
- routes/image_routes.py — Image upload, cover generation, frontmatter
- routes/publish_routes.py — Article publish, Git status, system publish
- routes/email_routes.py — Email push/preview
- routes/settings_routes.py — Settings CRUD (with runtime service rebuild)
- routes/socketio_routes.py — WebSocket event handlers

app.py reduced from 1119 to 247 lines. All 96 tests pass.

Each module follows the existing factory pattern from ai_routes.py:
a register_*_routes() function accepts dependencies via closure,
returns the Blueprint. Settings routes use a mutable settings_state
dict to support runtime service reconstruction (hugo_root change).

Closes #84
